+++ src/com/intellij/idea/plugin/hybris/flexibleSearch/psi/FlexibleSearchPsiUtil.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -21,13 +21,13 @@
 package com.intellij.idea.plugin.hybris.flexibleSearch.psi
 
 import com.intellij.idea.plugin.hybris.flexibleSearch.psi.impl.FlexibleSearchYColumnNameMixin
+import com.intellij.openapi.progress.ProgressManager
 import com.intellij.psi.PsiElement
 import com.intellij.psi.util.PsiTreeUtil
 import com.intellij.psi.util.childrenOfType
 import com.intellij.psi.util.parentOfType
 import com.intellij.psi.util.siblings
 import com.intellij.util.asSafely
-import org.jetbrains.plugins.groovy.lang.psi.util.backwardSiblings
 
 fun getPresentationText(resultColumn: FlexibleSearchResultColumn) = (
     resultColumn.childrenOfType<FlexibleSearchColumnAliasName>()
@@ -153,3 +153,8 @@ fun getTableAlias(element: FlexibleSearchDefinedTableName): FlexibleSearchTableA
     .siblings()
     .firstOrNull { it is FlexibleSearchTableAliasName }
     ?.asSafely<FlexibleSearchTableAliasName>()
+
+private fun PsiElement.backwardSiblings(): Sequence<PsiElement> = generateSequence(this) {
+    ProgressManager.checkCanceled()
+    it.prevSibling
+}